requirements.txt
fastapi
uvicorn
boto3
python-dotenv
jinja2
python-multipart